Moved some functions to apputil.cpp file
