+++ testing/srt-test-live.maf
@@ -3,6 +3,7 @@
 SOURCES
 srt-test-live.cpp
 testmedia.cpp
+../apps/apputil.cpp
 ../apps/verbose.cpp
 ../apps/socketoptions.cpp
 ../apps/uriparser.cpp
