+++ testing/srt-test-relay.maf
@@ -3,6 +3,7 @@
 SOURCES
 srt-test-relay.cpp
 testmedia.cpp
+../apps/apputil.cpp
 ../apps/verbose.cpp
 ../apps/socketoptions.cpp
 ../apps/uriparser.cpp
